Occasional Refactor
Going in-&-out of the creation process. I want to make sure to take the
time in-between to go back and refactor and ensure readability and
sustainability of the code just in case anyone wants to use the source.

I went through and updated variables and functions to explicitly state
their responsibility.
I also refactored the ALocalSimulationVolume::Update function and
following functions to lay ground work for future work.